refactor(progressView): remove metadata projection arms
