+++ src/shared/ipc.ts
@@ -128,9 +128,6 @@ export const PROGRESS_VIEW_COMMANDS = {
   SETTLE_STREAM_SELECTION: 'settleStreamSelection',
   RELEASE_STREAM_CONTENT: 'releaseStreamContent',
   UPDATE_CONVERSATION_PROGRESS: 'updateConversationProgress',
-  UPDATE_STAGE: 'updateStage',
-  UPDATE_STREAM_BADGES: 'updateStreamBadges',
-  UPDATE_STREAM_DESCRIPTION: 'updateStreamDescription',
   SYNC_INQUIRY_THREADS: 'syncInquiryThreads',
   UPDATE_INQUIRY_THREAD: 'updateInquiryThread',
   UPDATE_FILES: 'updateFiles',
